cmd: extract kill/--watch into cmd_device_kill.go
Pure move, no behavior change: runKillCommand, resolveBundleIDsToProcessNames,
runKillOnce, matchesKillTarget, runKillWatch, and their test all moved out
of cmd_device_apps.go into their own file, since kill/--watch had grown
into the largest, most self-contained chunk of that file.
