+++ cmd_device_kill_test.go
